refactor services and repositories for blogs and posts

+++ src/services/blogs-service.ts
@@ -3,25 +3,59 @@ import { ICreateBlogPayload } from '@/types/blogs/createBlogBody';
 import { IUpdateBlogPayload } from '@/types/blogs/updateBlogBody';
 import { ObjectId } from 'mongodb';
 import BlogsRepository from '@/repositories/blogs-repository';
+import { ErrorService } from './error-service';
+import { BlogsErrorsList } from '@/errors/blogs-errors';
+import { HTTP_STATUS_CODES } from '@/const/http-status-codes';
 
 const getAllBlogs = async (): Promise<IBlog[]> => {
   return await BlogsRepository.getAllBlogs();
 };
 
 const getBlogById = async (blogId: ObjectId): Promise<IBlog> => {
-  return await BlogsRepository.getBlogById(blogId);
+  if (!ObjectId.isValid(blogId)) {
+    throw ErrorService(BlogsErrorsList.NOT_FOUND, HTTP_STATUS_CODES.NOT_FOUND_404);
+  }
+
+  const blog = await BlogsRepository.getBlogById(blogId);
+
+  if (!blog) {
+    throw ErrorService(BlogsErrorsList.NOT_FOUND, HTTP_STATUS_CODES.NOT_FOUND_404);
+  }
+
+  return blog;
 };
 
 const createBlog = async (blog: ICreateBlogPayload): Promise<IBlog> => {
-  return await BlogsRepository.createBlog(blog);
+  const newBlog: IBlog = { ...blog, isMembership: false, createdAt: new Date() };
+  return await BlogsRepository.createBlog(newBlog);
 };
 
 const updateBlog = async (blogId: ObjectId, newBlog: IUpdateBlogPayload): Promise<void> => {
-  return await BlogsRepository.updateBlog(blogId, newBlog);
+  if (!ObjectId.isValid(blogId)) {
+    throw ErrorService(BlogsErrorsList.NOT_FOUND, HTTP_STATUS_CODES.NOT_FOUND_404);
+  }
+
+  const blogToUpdate = await BlogsRepository.getBlogById(blogId)
+
+  if (!blogToUpdate) {
+    throw ErrorService(BlogsErrorsList.NOT_FOUND, HTTP_STATUS_CODES.NOT_FOUND_404);
+  }
+
+  await BlogsRepository.updateBlog(blogId, newBlog);
 };
 
 const deleteBlog = async (blogId: ObjectId): Promise<void> => {
-  return await BlogsRepository.deleteBlog(blogId);
+  if (!ObjectId.isValid(blogId)) {
+    throw ErrorService(BlogsErrorsList.NOT_FOUND, HTTP_STATUS_CODES.NOT_FOUND_404);
+  }
+
+  const blogToDelete = await BlogsRepository.getBlogById(blogId)
+
+  if (!blogToDelete) {
+    throw ErrorService(BlogsErrorsList.NOT_FOUND, HTTP_STATUS_CODES.NOT_FOUND_404);
+  }
+
+  await BlogsRepository.deleteBlog(blogId);
 };
 
 export default {
